accumS should always be able to return a value, even if its upstream can't.
Prior to this commit, we would always abort the entire computation if we got
a Nothing somewhere in the call chain. However, consider the following:

    foo pred sig = accumS False (const True <$ filterS pred sig)

With this behaviour, foo would always return False, as filterS would not have
a value to pass on until after pred has become True. Since accumS has a
default value, it doesn't need its upstream to return a value. If upstream is
undefined, it's obviously not the update function, so we return the currently
accumulated value.